ModelAnimator setup (#8)
* ModelAnimator
This PR introduces ModelAnimator which will be ultimately used to compute all the bones to mimic the angles from the pose tracker. Currently, it doesn't do anything useful other than just turning Unity Chan to face her left (forward vector: (-1, 0, 0)).

* Joint
This also adds a Joint class which will be used to keep track of each joint's rotation. Eventually, the ModelAnimator will need to update all these joints to the right angles.

For now, we just play with the hip, but eventually it would need to compute all bones in the model.

* SampleScene
This PR adds a Unity-Chan model in the middle of the scene.