Add arm control. (#29)
Add control to the shoulder joint and elbow joint. The recognition is done in the new class PoseLandmarksRecognizer.

PoseLandmarksRecognizer will returns a list of PoseLandmarks joints, with each of it specify a Quaternion indicating how to rotate the joint locally to the desired pose.

Currently the code is a bit redundant since only four bones are related (shoulder * 2, elbow * 2). In the future if this needs to move more bones (e.g. lower body), this can be further expand to build a bone connection graph like in ModelAnimator.

Side effects to ModelAnimator:

The head rotation would not work. This is okay - since using pose landmarks for head rotation is less accurate anyway. We can add it back with the FaceLandmarksRecognizer in a follow-up.
A bonus for it: No need to require a nose Transform exists in the model.